+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Features
 
-- Improved model evaluation, autoselect metrics
+- Improved model evaluation
 - Improved preprocessing
 
 ## 0.1.1